test(prd-144): dedupe nonblank_rows/joined_rows into tests/common (Greptile P2)
The L1 buffer-render helpers `nonblank_rows` and `joined_rows` were copied
verbatim in both tests/render_layout.rs and tests/render_button_bar.rs.
Extract a single shared copy into tests/common/mod.rs (pub fn, new "L1
buffer-render helpers" section) and have both suites pull them via
`mod common;` — the same shared-helper mechanism the other integration
tests use. Pure test-side DRY refactor: no assertion, scenario, or
production code changed.

Addresses the non-blocking Greptile P2 finding on PR #166. The pre-existing
chars_next_cmp clippy --tests lint in tests/common/mod.rs (from PR #138) is
left as-is, out of scope.
